requirements.txt
flask
flask-cors
numpy
tensorflow==2.13.0
keras==2.13.1
pillow
gdown
gunicorn